- Use %zu when formatting size_t.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@13338 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/envvars/envvars_common.cpp
@@ -212,8 +212,8 @@ wxArrayString nsEnvVars::GetEnvvarsBySetPath(const wxString& set_path)
       return envvars;
 
     const wxArrayString envvars_keys(cfg->EnumerateKeys(set_path));
-    const unsigned long num_envvars = envvars_keys.GetCount();
-    for (unsigned long i = 0; i < num_envvars; ++i)
+    const size_t num_envvars = envvars_keys.GetCount();
+    for (size_t i = 0; i < num_envvars; ++i)
     {
         const wxString envvar(cfg->Read(set_path+'/'+envvars_keys[i]));
         if (!envvar.empty())
@@ -222,7 +222,7 @@ wxArrayString nsEnvVars::GetEnvvarsBySetPath(const wxString& set_path)
             EV_DBGLOG("Warning: empty envvar '%s' detected and skipped.", envvars_keys[i]);
     }
 
-    EV_DBGLOG("Read %zu/%lu envvars in path '%s'.", envvars.GetCount(), num_envvars, set_path);
+    EV_DBGLOG("Read %zu/%zu envvars in path '%s'.", envvars.GetCount(), num_envvars, set_path);
 
     return envvars;
 }// GetEnvvarsBySetPath
